fix: align integration tests with team standards
- Update jubilant to ~=1.8 and add pytest-jubilant>=2,<3
- Remove --use-existing and --model options from conftest.py (handled
  by pytest-jubilant via --juju-model, --no-juju-setup,
  --no-juju-teardown)
- --keep-models is now a no-op in tests/conftest.py for CI compatibility
- Simplify juju fixture to delegate lifecycle to pytest-jubilant,
  overriding only wait_timeout
- Move httprequest_lego_provider_image fixture to integration conftest.py
- Convert deployment from fixture to @pytest.mark.juju_setup test
- Use fixed unit name f"{APP_NAME}/0" instead of dynamic lookup
- Update copyright year to 2026 in modified files
