+++ charm/tests/conftest.py
@@ -1,4 +1,4 @@
-# Copyright 2025 Canonical Ltd.
+# Copyright 2026 Canonical Ltd.
 # See LICENSE file for licensing details.
 
 """Fixtures for charm tests."""
@@ -16,16 +16,5 @@ def pytest_addoption(parser):
         "--keep-models",
         action="store_true",
         default=False,
-        help="keep temporarily-created models",
-    )
-    parser.addoption(
-        "--use-existing",
-        action="store_true",
-        default=False,
-        help="use existing models and not created models",
-    )
-    parser.addoption(
-        "--model",
-        action="store",
-        help="temporarily-created model name",
+        help="Keep models after tests (no-op, for CI compatibility).",
     )
