+++ tox.ini
@@ -48,7 +48,7 @@ deps =
     flake8-docstrings-complete>=1.0.3
     flake8-test-docs>=1.0
     isort
-    jubilant~=1.7
+    jubilant~=1.8
     mypy
     pep8-naming
     pydocstyle>=2.10
@@ -95,8 +95,9 @@ commands =
 description = Run integration tests (placeholder)
 deps =
     -r{toxinidir}/charm/requirements.txt
-    jubilant~=1.7
+    jubilant~=1.8
     pytest
+    pytest-jubilant>=2,<3
 commands =
     pytest -v --tb native \
         --log-cli-level=INFO -s {posargs} charm/tests/integration
